refactor(tokens): removing tokens/date-picker
