+++ packages/components/date-picker/src/composables/use-range-picker.ts
@@ -1,8 +1,8 @@
 import { getCurrentInstance, inject, ref, unref, watch } from 'vue'
 import { isArray } from '@element-plus/utils'
-import { ROOT_PICKER_INJECTION_KEY } from '@element-plus/tokens'
 import { useLocale, useNamespace } from '@element-plus/hooks'
 import { getDefaultValue, isValidRange } from '../utils'
+import { ROOT_PICKER_INJECTION_KEY } from '../constants'
 import { useShortcut } from './use-shortcut'
 
 import type { Ref } from 'vue'
