+++ packages/tokens/index.ts
@@ -1,2 +1,1 @@
 export * from './experimental-features'
-export * from './date-picker'
